requirements.txt
selenium
pytest
pytest-xdist
allure-pytest
webdriver-manager
python-dotenv
loguru